[LINT] error 429 (bis)
Custodial pointer 'Symbol' (Location) has not been freed or returned.

I've split this into it own commit as the changes are more heavy and might
not be agreed to by Dave and therefore can be cherry-picked out should he
wants to do without.

I've modified the way the PyProducerConsumer works to deal with shared_ptrs
instead of plain old pointers. This has the clear advantage of freeing the
developer(s) of the burden of memory allocation/deallocation. Although it
appears memory was dealt with properly in the context of the
PyProducerConsumers, the code paths between allocation and deallocation wasn't
clear at first glance (memory allocated before call to produce and deallocated
in the consume call) and was error prone. The solution implemented here
addresses this problem.

Signed-off-by: Jocelyn Legault <jocelynlegault@gmail.com>