Debugging - do not ignore release request from thread that allegedly does not have GIL
After lots of debugging, the issue seems to be the exact race condition mentioned in the comments.
The variable is set or reset out of order with the actual GIL acquisition.

Not having this in will cause an issue in certain other circumstances, so it's only a temporary hack
to prove the issue is just the ordering of the GIL lock/release and resetting the thread ID with the GIL.

Need to understand what order things need to happen in, in order for this to be wrong.