CallbackExecArgs deallocation bug
CallbackExecArgs contains boost python objects, and was deallocated with a
shared_ptr. Unfortunately this wasn't always when we still had the GIL.
Changed to pointers, then we explicitly grab the GIL when in the destructor.

CallbackExecArgs is now in the namespace NppPythonScript - many other classes
also removed from global namespace.  There's more to do in this area though.