Update intro.rst, Handling Notifications: the Callback smallprint
A re-write of the sub-section "The callback smallprint". 
The original was very terse and hard to undertand for people who don't have a computer science degree (or  experience with events processing). See https://github.com/bruderstein/PythonScript/issues/298

It improves the explanations and also adds 2 example scripts. 
Since it got a bit bigger, I changed the title as well to match the content: "Synchronous and asynchronous callbacks".